+++ views/legend.php
@@ -3,7 +3,7 @@ namespace Fragen\Category_Colors;
 
 use Tribe__Events__Main;
 
-$teccc = Main::instance();
+$teccc->setup_terms( $options );
 
 ?>
 <div id="legend_box" class="teccc-legend">
@@ -24,7 +24,6 @@ $teccc = Main::instance();
 			</li>
 		<?php endforeach ?>
 
-		<?php $options = get_option( 'teccc_options' ); ?>
 		<?php if ( isset( $options['show_ignored_cats_legend'] ) ) : ?>
 			<?php foreach ( $teccc->ignored_terms as $ignored_term ) : ?>
 				<?php